pre-demo snapshot
