--- Stock-Information/requirements.txt
@@ -1,10 +1,0 @@
-
-firebase_admin
-flask
-flask_cors
-matplotlib
-numpy
-pandas
-pathlib
-ta
-yfinance
