+++ package.json
@@ -12,6 +12,7 @@
     "@testing-library/user-event": "^7.2.1",
     "core-js": "^2.6.11",
     "firebase": "^7.16.0",
+    "npm-install-missing": "^0.1.4",
     "react": "^16.13.1",
     "react-bootstrap": "^1.2.2",
     "react-dom": "^16.13.1",
@@ -19,6 +20,7 @@
     "react-hot-loader": "^4.12.21",
     "react-native": "^0.63.2",
     "react-native-web": "^0.13.4",
+    "react-numeric-input": "^2.2.3",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^3.4.1",
