+++ src/components/authenticate/UserProvider.jsx
@@ -1,5 +1,5 @@
 import React, { Component, createContext } from "react";
-import { auth } from "../../../firebase";
+import { auth } from "../../firebase";
 export const UserContext = createContext({ user: null });
