+++ src/components/firebase-access.jsx
@@ -10,10 +10,12 @@ export const addUser = async (roomID, nickname) => {
     const numDays = (await getDates(roomID)).length;
     const numSymbols = (await getSymbols(roomID)).length;
     const roomRef = await getRoomRef(roomID);
-    const userRef = await roomRef.collection('users').doc();
+    const userRef = roomRef.collection('users').doc();
     const userID = userRef.id;
     const empArray = Array.from(Array(numSymbols), () => 0);
     const startingMoney = await getStartingMoney(roomID);
+
+    console.log(userRef);
 
     const gameInfo = {
         userId: userID,
@@ -22,6 +24,7 @@ export const addUser = async (roomID, nickname) => {
         money_left: startingMoney,
         curShares: empArray,
     }
+    console.log(gameInfo);
 
     userRef.set(gameInfo);
     for(var i_day = 0; i_day < numDays; i_day++) {
@@ -76,8 +79,8 @@ export const changeShares = async (roomID, userID, dayIndex, changeArray) => {
 }
 
 export const initDates = async (symbols, Rounds) => {
-    let Stocks = await db.collection("Ticker-Info").doc("Stock").collection("Stocks")
-        .where("Symbol","in", symbols).get();
+    let Stocks= await db.collection("Ticker-Info").doc("Stock").collection("Stocks")
+        .where("Symbol","in",symbols).get()
     let IPOyearMax = 0;
     let today = new Date();
     let year = today.getFullYear();
@@ -89,19 +92,19 @@ export const initDates = async (symbols, Rounds) => {
 
     // No more than 7 rounds(Periods are measured in months)
     let min_window_size = 3;
-    let yearDiff = year - (IPOyearMax + 1);
+    let yearDiff = year - (IPOyearMax+1);
     let maximum_period = Math.floor(((yearDiff * 12)  - min_window_size) / Rounds);
-    let random_period =  Math.floor((Math.random()  * (maximum_period - min_window_size)) + min_window_size);
-    let startDate = new Date(IPOyearMax + 1, 1, 1);
-    let endDate =  new Date(IPOyearMax + 1, 1 + random_period, 1);
-    let rand_startDate = randomDate(startDate,endDate);
+    let random_period =  Math.floor((Math.random()  * (maximum_period - min_window_size))+min_window_size);
+    let startDate = new Date(IPOyearMax+1,1,1);
+    let endDate =  new Date(IPOyearMax+1,1+random_period,1);
+    let rand_startDate =  randomDate(startDate,endDate);
     let dates = [];
     let curr_date = rand_startDate;
     for(var i = 0; i < Rounds; i++) {
         dates.push(curr_date.toISOString().substring(0, 10));
-        curr_date = new Date(curr_date.setMonth(curr_date.getMonth() + random_period));
+        curr_date = new Date(curr_date.setMonth(curr_date.getMonth()+random_period));
     }
-    const datesD = {
+    const datesD ={
         "dates" : dates,
         "period": random_period
     }
@@ -111,14 +114,14 @@ export const initDates = async (symbols, Rounds) => {
 export const initializeQuiz = async (symbols, roomId, periodLen, endDates) => {
     var formData = new FormData();
     formData.append('symbol',JSON.stringify(symbols));
-    formData.append('RoomId', roomId);
+    formData.append('RoomId',roomId);
     formData.append('end-date',JSON.stringify(endDates));
     try{
         await fetch('http://localhost:8080/get-prices', {
             method: 'POST',
             mode: 'cors',
             body: formData
-        });
+        })
     }
     catch(err) {
         console.log("Error is " +  err)
@@ -136,18 +139,13 @@ export const initializeQuiz = async (symbols, roomId, periodLen, endDates) => {
     }
 }
 
-export const initSymbols = async(db,Industry,Sector,MarketCap,NumOfSymbols) =>{
-    let symbols = [];
-    if(atLeastTwo(Industry,Sector,MarketCap)){
-        //In the case where a person wants to query using multiple features it may return a larger request,
-        //so we send it to the backend
+export const initSymbols = async(Industry,Sector,NumOfSymbols) =>{
+    let symbols = []
+    if (Sector !== null && Industry !== null){
+
         let formData = new FormData();
-        if(Industry)
-            formData.append('Industry',Industry);
-        if(Sector)
-            formData.append('Sector',Sector);
-        if(MarketCap)
-            formData.append('MarketCap',MarketCap);
+        formData.append('Industry',Industry);
+        formData.append('Sector',Sector);
         formData.append('NumOfSymbols',NumOfSymbols);
         try{
             let response = await fetch('http://localhost:8080/get-symbols', {
@@ -155,41 +153,31 @@ export const initSymbols = async(db,Industry,Sector,MarketCap,NumOfSymbols) =>{
                 mode: 'cors',
                 body: formData
             })
-            let symbolJson = await response.json();
+            let symbolJson = await response.json()
             if (symbolJson.hasOwnProperty("Error")){
                 console.log("No Symbols for your query")
-                return symbols;
+                return symbols
             }
-            symbols = symbolJson['symbols'];
+            symbols = symbolJson['symbols']
         }
         catch(error){
-            console.log("Error with Query: " + error);
+            console.log("Error with Query: " + error)
         }
+
     }
-    // All of the rest of the values use a search by Xpos to make the query as small as possible(O(numOfSymbols))
     else if(Industry !== null){
+
         let IndustryInfo =  await db.collection("Ticker-Info").doc("Industry").get();
         let numOfIndustries= IndustryInfo.data().Industry[Industry];
         let cutoff = Math.floor((Math.random()  * (numOfIndustries - NumOfSymbols))+NumOfSymbols);
         let Industries = await db.collection("Ticker-Info").doc("Stock").collection("Stocks")
             .where("Industry","==",Industry)
             .where("IndustryPos","<=", cutoff)
-            .orderBy("IndustryPos").limit(NumOfSymbols).get();
+            .orderBy("IndustryPos").limit(NumOfSymbols).get()
         Industries.forEach(function(doc){
-            symbols.push(doc.data().Symbol);
+            symbols.push(doc.data().Symbol)
         })
-    }
-    else if(MarketCap !== null){
-        let MarketCapInfo = await db.collection("Ticker-Info").doc("Market-Cap").get();
-        let numStocks= MarketCapInfo.data().MarketCap[MarketCap];
-        let cutoff = Math.floor((Math.random()  * (numStocks - NumOfSymbols))+NumOfSymbols);
-        let Stocks = await db.collection("Ticker-Info").doc("Stock").collection("Stocks")
-            .where("MarketCapSize","==",MarketCap)
-            .where("MarketCapPos","<=", cutoff)
-            .orderBy("MarketCapPos").limit(NumOfSymbols).get();
-        Stocks.forEach(function(doc){
-            symbols.push(doc.data().Symbol);
-        })
+
     }
     else if(Sector !== null){
 
@@ -199,10 +187,11 @@ export const initSymbols = async(db,Industry,Sector,MarketCap,NumOfSymbols) =>{
         let Sectors = await db.collection("Ticker-Info").doc("Stock").collection("Stocks")
             .where("Sector","==",Sector)
             .where("SectorPos","<=", cutoff)
-            .orderBy("SectorPos").limit(NumOfSymbols).get();
+            .orderBy("SectorPos").limit(NumOfSymbols).get()
         Sectors.forEach(function(doc){
-            symbols.push(doc.data().Symbol);
-        });
+            symbols.push(doc.data().Symbol)
+        })
+
     }
     else{
         let StockInfo =  await db.collection("Ticker-Info").doc("Stock").get();
@@ -210,12 +199,13 @@ export const initSymbols = async(db,Industry,Sector,MarketCap,NumOfSymbols) =>{
         let cutoff = Math.floor((Math.random()  * (numOfStocks - NumOfSymbols))+NumOfSymbols);
         let Stocks = await db.collection("Ticker-Info").doc("Stock").collection("Stocks")
             .where("RandomPos",">=", cutoff)
-            .orderBy("RandomPos").limit(NumOfSymbols).get();
+            .orderBy("RandomPos").limit(NumOfSymbols).get()
         Stocks.forEach(function(Stock){
-            symbols.push(Stock.data().Symbol);
-        });
+            symbols.push(Stock.data().Symbol)
+        })
+
     }
-    return symbols;
+    return symbols
 }
 
 export const makeInvestment = async (roomID, userID, dayIndex, changeArray) => {
@@ -234,21 +224,24 @@ function randomDate(start, end) {
     return date;
 }
 
-export const setUpRoom = (symbolsL, Rounds, password, startingMoney = 10000) => {
+export const setUpRoom = (numOfSymbols,rounds,password,startingMoney = 10000) => {
+
     const roomRef = db.collection('Rooms').doc();
     roomRef.set({
         day_index: 0,
         phase: 'no-host',
         password: password,
-        startingMoney: startingMoney,
+        starting_money: startingMoney,
     });
     const roomID = roomRef.id;
-    initDates(db,symbolsL,Rounds).then((datesD)=> {
-        roomRef.update({
-            symbols: symbolsL,
-            dates: datesD["dates"],
+    initSymbols(null,null,numOfSymbols).then((symbolsL) => {
+        initDates(symbolsL,rounds).then((datesD)=> {
+            roomRef.update({
+                symbols: symbolsL,
+                dates: datesD["dates"],
+            });
+            initializeQuiz(symbolsL,roomID,datesD["period"],datesD["dates"]);
         });
-        initializeQuiz(symbolsL,roomID,datesD["period"],datesD["dates"]);
     });
     return roomID;
 }
@@ -372,7 +365,7 @@ export const getShares = async (roomID, userID) => {
 
 export const getStartingMoney = async (roomID) => {
     const roomData = await getRoomData(roomID);
-    return roomData.startingMoney;
+    return roomData.starting_money;
 }
 
 // retrieves symbol name given the symbol's index
@@ -395,51 +388,4 @@ export const getUserData = async (roomID, userID) => {
 
 export const getUserRef = (roomID, userID) => {
     return db.collection('Rooms').doc(roomID).collection('users').doc(userID);
-}
-
-/* configuration methods */
-function compDoc(a, b){
-  if (a.value < b.value) {
-    return -1;
-  }
-  if (a.value > b.value) {
-    return 1;
-  }
-  return 0;
-}
-
-export const getIndustries = async (db) => {
-  let Industries = await db.collection("Ticker-Info").doc("Industry").get();
-  let IndustryData =  Industries.data().Industry;
-  let IndustryList = [];
-  for(const Industry in IndustryData) {
-    IndustryList.push({value: Industry, label: Industry});
-  }
-  IndustryList.push({value: null, label: "None"});
-  IndustryList.sort(compDoc);
-  return IndustryList;
-}
-
-export const getSectors = async (db) => {
-  let Sectors = await db.collection("Ticker-Info").doc("Sector").get();
-  let SectorData =  Sectors.data().Sector;
-  let SectorList = [];
-  for(const Sector in SectorData) {
-    SectorList.push({value: Sector, label: Sector});
-  }
-  SectorList.push({value: null, label:"None"});
-  SectorList.sort(compDoc);
-  return SectorList;
-}
-
-export const getMarketCaps = async (db) => {
-  let MarketCaps = await db.collection("Ticker-Info").doc("Market-Cap").get();
-  let MarketCapData =  MarketCaps.data().MarketCap;
-  let MarketCapList = [];
-  for(const Stock in MarketCapData) {
-    MarketCapList.push({value: Stock, label: Stock});
-  }
-  MarketCapList.push({value: null, label:"None"});
-  MarketCapList.sort(compDoc);
-  return MarketCapList;
 }
