+++ src/components/quiz/Instructions.jsx
@@ -14,6 +14,7 @@ class Instructions extends Component {
             - 'play' -> for players to read (in play page)
             - 'create' -> for creating the room
      */
+
     constructor(props) {
         super(props);
         this.state = {
@@ -51,7 +52,30 @@ class Instructions extends Component {
                         {info === 'play' &&
                             <div>
                                 <h1>This is the play page</h1>
-                                <p> After you join a game, you can ASDF</p>
+                                <h2>There are a couple of things you can see and can do to invest</h2>
+                                <h3>You can see</h3>
+                                <p>- Your current net worth, which is your cash + the value of your shares</p>
+                                <p>- Your current cash, which you can buy shares with</p>
+                                <p>After you select a symbol from the dropdown, you can see</p>
+                                <p>- How many shares of that symbol you have</p>
+                                <p>- How much each share costs</p>
+                                <p>- Graphs that have multiple technical indicators for the symbol</p>
+                                <p>- A box to enter the number of shares you want to buy or sell</p>
+                                <p>- The option to buy or sell a certain number of shares</p>
+                                <h2> Progression </h2>
+                                <p>The way time works in this game is that the host has predefined "stages"</p>
+                                <p>In each of these stages, you can buy or sell shares at a given price</p>
+                                <p>Whoever has the largest net worth after the last stage wins</p>
+                                <h2> Stage Control </h2>
+                                <p> For now, stage control is entirely decided by the host. So once they click "next stage"
+                                    it will be reflected across all players</p>
+                            </div>
+                        }
+
+                        {info === 'ta' &&
+                            <div>
+                                <h1> you are looking at the technical analysis indicators </h1>
+                                <h2> more explanations to be added soon </h2>
                             </div>
                         }
