+++ src/components/styles/Instructions.jsx
