+++ src/components/styles/Play.jsx
@@ -1,0 +1,9 @@
+import Background from "../assets/background.jpg";
+
+export const body = {
+    fontFamily: "Calibri",
+    textAlign: "center",
+    backgroundImage: "url("+Background+")",
+    minHeight: '105vh',
+    minWidth: '100%',
+}
