+++ stock_backend/.gitignore
