+++ stock_backend/requirements.txt
@@ -1,0 +1,23 @@
+yfinance
+numpy
+flask
+flask_cors
+firebase_admin
+pathlib
+pandas
+grcpio
+protobuf
+wrapt
+google-auth<2,>=1.6.3
+werkzeug
+keyring
+pylint
+colorama
+Jinja2
+packaging
+oautlib
+cffi
+cryptography
+jsonschema!=2.5.0
+colorama
+entrypoints
