+++ stock_backend/stock_data.py
@@ -21,9 +21,10 @@ firebase_admin.initialize_app(cred,{
 firebase_request_adapter = requests.Request()
 
 app = Flask(__name__)
-app.secret_key = ""
+app.secret_key = "" # Secret key can't be on github
 app.config['SESSION_TYPE'] = 'filesystem'
 app.config['IMAGES'] = 'images/'
+CORS(app)
 db = firestore.client()
 bucket = storage.bucket()
 
@@ -86,32 +87,21 @@ def get_stock_image():
                     'Stockpublic_image_url':StockDict,
             }
             db.collection('Rooms').document(roomID).collection(symbol).document('images').set(img)
-    response = app.response_class(
-        response=json.dumps({'success':True}),
-        status=200,
-        mimetype='application/json'
-    )
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response 
+    return json.dumps({'success':True}), 200, {'ContentType':'application/json'} 
 
 @app.route('/get-symbols',methods=['POST'])
 def get_symbols():
-    Num_of_Symbols = int(request.form['NumOfSymbols'])
-    StocksQuery = db.collection("Ticker-Info").document("Stock").collection("Stocks")
-    if 'Industry' in request.form:
-        print(request.form['Industry'])
-        StocksQuery = StocksQuery.where("Industry","==",request.form['Industry'])
-    if 'Sector'  in request.form:
-        print(request.form['Sector'])
-        StocksQuery = StocksQuery.where("Sector","==",request.form['Sector'])
-    if 'MarketCap' in request.form:
-        print(request.form['MarketCap'])
-        StocksQuery = StocksQuery.where("MarketCapSize","==",request.form['MarketCap'])
+    Industry = request.form['Industry']
+    Sector = request.form['Sector']
+    Num_of_Symbols = request.form['NumOfSymbols']
+
+    IndSect = db.collection("Ticker-Info").doc("Stock").collection("Stocks").where("Industry","==",Industry) \
+    .where("Sector","==", Sector).stream()
     num_of_stocks = 0
     symbols = []
-    for stock in StocksQuery.stream():
+    for stock in IndSect:
         num_of_stocks+=1
-        symbols.append(stock.to_dict()['Symbol'])
+        symbols.append(stock.to_dict()['Stock Data']['Symbol'])
 
     responseD ={}
     if num_of_stocks == 0:
@@ -126,12 +116,12 @@ def get_symbols():
         responseD = {
             "symbols": random.choices(symbols,k=Num_of_Symbols)
         }
+
     response = app.response_class(
         response=json.dumps(responseD),
         status=200,
         mimetype='application/json'
     )
-    response.headers.add('Access-Control-Allow-Origin', '*')
     return response
     
 
@@ -154,13 +144,7 @@ def tech_indic():
     TechI = IntrinsicValue.getAllTechnicalIndicators(symbol,end_date,period)
     TechDict = TechI.to_dict(orient='index')
     db.collection('Rooms').document(roomID).collection(symbol).document('technical-indicators').set(TechDict)
-    response = app.response_class(
-        response=json.dumps({'success':True}),
-        status=200,
-        mimetype='application/json'
-    )
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
+    return json.dumps({'success':True}), 200, {'ContentType':'application/json'} 
 
 '''
 Params:
@@ -188,13 +172,7 @@ def time_series():
             'prices': prices
         }
         db.collection('Rooms').document(roomID).collection(symbol).document('Prices').set(price)
-    response = app.response_class(
-        response=json.dumps({'success':True}),
-        status=200,
-        mimetype='application/json'
-    )
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
+    return json.dumps({'success':True}), 200, {'ContentType':'application/json'} 
      
 if __name__ == '__main__':
     # This is used when running locally only. When deploying to Google App
